Physics Fix
- Default physics object state is disabled, this is like this so on duplicate scene on play, the duplicated scene's one dont affect the current scene ones
- Disabled pixel to meters in physics, they were already calculated in other part of the code
- Recreate body function in case m_body is nullptr
- Added bounce coefficient
- Improved GUI: on exit it will ask for save, dont save, etc, and if that savescenepanel you want to cancel, on kittinc escape key it will close

- All physics seems to be nice except for ->
- TODO: first time to hit play, they wont move, if stop and play again they will, find why
- TODO: find why physics_scene.mlscene freezes the program on loading

+++ ManLite/ManLiteEngine/PhysicsEM.h
@@ -32,7 +32,7 @@ public:
 
 	bool Awake();
 	bool Start();
-    bool Update(float dt);
+    bool Update(double dt);
 	bool CleanUp();
 
 	//
